SettingElement, IAmSettable, and SettingNames init.

+++ Assets/_THALASSO/Coding/Scripts/Audio/WwiseHelperScripts/WwiseBus.cs
@@ -7,80 +7,87 @@ namespace WwiseHelper
     {
 #if WWISE_2024_OR_LATER
         private float _volume = 0.8f;
-        private float _volumeRange = 1.0f;
-
 
         private readonly AK.Wwise.RTPC _rtpc;
         private readonly Slider _volumeSlider;
 
-
-        public WwiseBus(AK.Wwise.RTPC rtpc, Slider volumeSlider)
+        public WwiseBus(AK.Wwise.RTPC rtpc, Slider slider)
         {
             _rtpc = rtpc;
-            _volumeSlider = volumeSlider;
+            _volumeSlider = slider;
 
             Volume = 0.8f; // set Volume at 80%
         }
+
+        public string RTPCName => _rtpc.Name;
 
         public float Volume
         {
             get => _volume;
             private set
             {
-                if (value != _volume)
-                {
-                    value /= _volumeRange;
+                if (value >= 1.0f)
+                    value = 1.0f;
 
-                    if (value >= 1.0f)
-                        value = 1.0f;
+                if (value <= 0.0f)
+                    value = 0.0f;
 
-                    if (value <= 0.0f)
-                        value = 0.0f;
+                _volume = value;
 
-                    _volume = value;
-                }
+                AkUnitySoundEngine.SetRTPCValue(_rtpc.Id, _volume);
             }
         }
 
+        #region Callback Registration/Deregistration
         public void AddListener()
         {
+            GlobalEventBus.Register(GlobalEvents.UI.MenuClosed, OnMenuClosed);
             _volumeSlider.onValueChanged.AddListener(SetVolume);
         }
 
-        public void DeleteData()
-        {
-            PlayerPrefs.DeleteKey(_rtpc.Name);
-        }
-
-        public void LoadVolume()
-        {
-            if (!PlayerPrefs.HasKey(_rtpc.Name))
-                PlayerPrefs.SetFloat(_rtpc.Name, Volume);
-
-            _volumeSlider.value = PlayerPrefs.GetFloat(_rtpc.Name);
-        }
 
         public void RemoveListener()
         {
             _volumeSlider.onValueChanged.RemoveListener(SetVolume);
+            GlobalEventBus.Deregister(GlobalEvents.UI.MenuClosed, OnMenuClosed);
         }
+        #endregion
 
-        public void SetVolumeRange(float minValue, float maxValue)
+        #region Public Methods
+        public void LoadData()
         {
-            _volumeRange = maxValue - minValue;
+            if (!PlayerPrefs.HasKey(_rtpc.Name))
+                PlayerPrefs.SetFloat(_rtpc.Name, Volume);
+
+            Volume = PlayerPrefs.GetFloat(_rtpc.Name);
+
+            _volumeSlider.value = _volumeSlider.maxValue * Volume + _volumeSlider.minValue * (1 - Volume);
         }
 
-        private void SaveVolume()
+        public void DeleteData()
+        {
+            if (PlayerPrefs.HasKey(_rtpc.Name))
+                PlayerPrefs.DeleteKey(_rtpc.Name);
+        }
+        #endregion
+
+        #region Private Methods
+        private void OnMenuClosed(object[] eventArgs)
+        {
+            SaveData();
+        }
+
+        private void SaveData()
         {
             PlayerPrefs.SetFloat(_rtpc.Name, Volume);
         }
 
         private void SetVolume(float volume)
         {
-            Volume = volume;
-            AkUnitySoundEngine.SetRTPCValue(_rtpc.Id, Volume);
-            SaveVolume();
+            if (_volumeSlider.normalizedValue != Volume)
+                Volume = _volumeSlider.normalizedValue;
         }
+        #endregion
 #endif
     }
 }
